requirements.txt
gitdb==4.0.7
GitPython==3.1.17
smmap==4.0.0
typing-extensions==3.10.0.0
